fix: restructure repo
