+++ megaparse/core/tests/pdf/test_pdf_processing.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from uuid import uuid4
 
 import pytest
