--- megaparse/parser/__init__.py
@@ -1,4 +1,0 @@
-from megaparse.parser.megaparser import MegaParser
-
-
-__all__ = ["MegaParser"]
